ai-enricher/requirements.txt
kafka-python
python-snappy
requests